apps/api/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis & Celery
redis==5.0.1
celery==5.3.4

# AWS S3
boto3==1.29.7
botocore==1.32.7

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1


